requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
python-telegram-bot==21.10
torch==2.6.0+cpu
sentence-transformers==3.4.1
faiss-cpu==1.10.0
numpy==2.2.4
Flask==3.0.3
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn[standard]==0.30.0
openai==1.35.0
pydantic==2.9.0
httpx==0.27.0
